+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 
 - 📈(frontend) track new recording's modes
 - ♿️(frontend) improve SR and focus for transcript and recording #810
+- 💄(frontend) adjust spacing in the recording side panels
 - 🚸(frontend) remove the default comma delimiter in humanized durations
 
 ### Fixed